requirements.txt
asgiref==3.3.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
Django==3.0.3
django-cors-headers==3.2.1
django-crontab==0.7.1
django-filter==2.4.0
django-oauth-toolkit==1.4.1
django-ranged-response==0.2.0
django-softdelete==0.9.2
django-xff==1.3.0
djangorestframework==3.11.0
django-rest-framework-proxy==1.6.0
django-utils==0.0.2
django-utils-six==2.0
django-simple-captcha==0.5.13
django-import-export==2.4.0
idna==2.10
lxml==4.6.2
M2Crypto==0.35.2
mysqlclient==1.4.6
oauthlib==3.1.0
pycparser==2.20
pyOpenSSL==19.1.0
python-docx==0.8.10
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
xlwt==1.3.0
pycryptodome== 3.9.9
uwsgi==2.0.19.1